[STYLE] Refactored new test-code to solve a cyclic import regression as per #377
* Also resolves multiple quality blockers from automated review of draft PR #376
