+++ tests/_test_documentation_status.py
@@ -36,7 +36,7 @@ To be documented?
 from . import (
     unittest,
 )
-from .test_readme_examples import (
+from ._readme_utils import (
     discover_examples,
 )
